requirements.txt
torch>=2.1.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
openai>=1.3.0
faiss-cpu>=1.7.4
langchain>=0.0.340
langchain-openai>=0.0.2
langchain-community>=0.0.10
tiktoken>=0.5.1
sentence-transformers>=2.2.2
plotly>=5.17.0
python-dotenv>=1.0.0
httpx>=0.25.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pydantic>=2.4.0
requests>=2.31.0
biopython>=1.81
joblib>=1.3.0
